+++ CHANGELOG.md
@@ -4,9 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Note - after update to 3.15.0 need remove the accessory frome Home app and add it again.
+## Note - after update to 3.15.x need remove the accessory frome Home app and add it again.
 
-## [3.15.7] - (30.12.2021)
+## [3.15.9] - (31.12.2021)
+### Fixed
+- power state report if device is disconnected from network
+
+### Changs
+- code cleanup
+
+## [3.15.7/8] - (30.12.2021)
 ### Changs
 - log AVR Disconnected if plugin restart and AVR not on network